+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/DownloadsFeature.kt
@@ -33,6 +33,7 @@ import mozilla.components.feature.downloads.manager.onDownloadStopped
 import mozilla.components.feature.downloads.ui.DownloaderApp
 import mozilla.components.feature.downloads.ui.DownloadAppChooserDialog
 import mozilla.components.lib.state.ext.flowScoped
+import mozilla.components.support.base.dialog.DeniedPermissionDialogFragment
 import mozilla.components.support.base.feature.LifecycleAwareFeature
 import mozilla.components.support.base.feature.OnNeedToRequestPermissions
 import mozilla.components.support.base.feature.PermissionsFeature
@@ -213,6 +214,7 @@ class DownloadsFeature(
             } else {
                 closeDownloadResponse(tab.id)
                 useCases.consumeDownload(tab.id, download.id)
+                showPermissionDeniedDialog()
             }
         }
     }
@@ -394,6 +396,16 @@ class DownloadsFeature(
         val positiveButtonTextColor: Int? = null,
         val positiveButtonRadius: Float? = null
     )
+
+    @VisibleForTesting
+    internal fun showPermissionDeniedDialog() {
+        fragmentManager?.let {
+            val dialog = DeniedPermissionDialogFragment.newInstance(
+                R.string.mozac_feature_downloads_write_external_storage_permissions_needed_message
+            )
+            dialog.showNow(fragmentManager, DeniedPermissionDialogFragment.FRAGMENT_TAG)
+        }
+    }
 }
 
 @VisibleForTesting