+++ mobile/android/android-components/components/feature/downloads/src/test/java/mozilla/components/feature/downloads/DownloadsFeatureTest.kt
@@ -403,6 +403,7 @@ class DownloadsFeatureTest {
 
         verify(downloadManager, never()).download(any(), anyString())
         verify(feature).closeDownloadResponse("test-tab")
+        verify(feature).showPermissionDeniedDialog()
     }
 
     @Test