+++ mobile/android/android-components/components/feature/prompts/src/test/java/mozilla/components/feature/prompts/dialog/ConfirmDialogFragmentTest.kt
@@ -8,7 +8,7 @@ import android.content.DialogInterface
 import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import mozilla.ext.appCompatContext
+import mozilla.components.support.test.ext.appCompatContext
 import org.junit.Assert
 import org.junit.Assert.assertEquals
 import org.junit.Before