+++ mobile/android/android-components/components/support/test/src/main/java/mozilla/components/support/test/ext/Context.kt
@@ -2,9 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.ext
+package mozilla.components.support.test.ext
 
 import android.content.Context
+import androidx.annotation.VisibleForTesting
 import androidx.appcompat.R
 import androidx.appcompat.view.ContextThemeWrapper
 import mozilla.components.support.test.robolectric.testContext
@@ -14,5 +15,5 @@ import mozilla.components.support.test.robolectric.testContext
  *
  * Useful for views that uses theme attributes, for example.
  */
-internal val appCompatContext: Context
+@VisibleForTesting val appCompatContext: Context
     get() = ContextThemeWrapper(testContext, R.style.Theme_AppCompat)